refactor(server): #125 — extract connect/forward handlers into routes/connect.rs
